[HW5] [v4]: modify func arg type in all kernel.cu

+++ HW5/src/kernel2.cu
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-__global__ void mandelKernel(int *device_data, int lowerX, int lowerY, int stepX, int stepY, size_t pitch, int maxIterations){
+__global__ void mandelKernel(int *device_data, float lowerX, float lowerY, float stepX, float stepY, size_t pitch, int maxIterations){
     // To avoid error caused by the floating number, use the following pseudo code
     //
     // float x = lowerX + thisX * stepX;
